Added utility class that memoizes the factorials previously computed
This class should be especially useful when many factorials are needed. The optimal way to use it is to first compute the biggest factorial; from there on, it's just a matter of looking at the table.
Also added some tests and a little benchmark to confirm it is worth it to compute some factorials using this class.